Meet 5 Done

+++ database/migrations/2023_08_23_131853_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->decimal('price');
+            $table->double('price');
             $table->string('image_url');
             $table->foreignIdFor(Category::class);
             $table->foreignIdFor(User::class);
